modify around icons
