--- src/utils/svg.jsx
@@ -1,73 +1,0 @@
-// Tech SNS SVG
-export const Github = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 28"><path d="M12 2c6.625 0 12 5.375 12 12 0 5.297-3.437 9.797-8.203 11.391-.609.109-.828-.266-.828-.578 0-.391.016-1.687.016-3.297 0-1.125-.375-1.844-.812-2.219 2.672-.297 5.484-1.313 5.484-5.922 0-1.313-.469-2.375-1.234-3.219.125-.313.531-1.531-.125-3.187-1-.313-3.297 1.234-3.297 1.234a11.28 11.28 0 0 0-6 0S6.704 6.656 5.704 6.969c-.656 1.656-.25 2.875-.125 3.187-.766.844-1.234 1.906-1.234 3.219 0 4.594 2.797 5.625 5.469 5.922-.344.313-.656.844-.766 1.609-.688.313-2.438.844-3.484-1-.656-1.141-1.844-1.234-1.844-1.234-1.172-.016-.078.734-.078.734.781.359 1.328 1.75 1.328 1.75.703 2.141 4.047 1.422 4.047 1.422 0 1 .016 1.937.016 2.234 0 .313-.219.688-.828.578C3.439 23.796.002 19.296.002 13.999c0-6.625 5.375-12 12-12zM4.547 19.234c.031-.063-.016-.141-.109-.187-.094-.031-.172-.016-.203.031-.031.063.016.141.109.187.078.047.172.031.203-.031zm.484.532c.063-.047.047-.156-.031-.25-.078-.078-.187-.109-.25-.047-.063.047-.047.156.031.25.078.078.187.109.25.047zm.469.703c.078-.063.078-.187 0-.297-.063-.109-.187-.156-.266-.094-.078.047-.078.172 0 .281s.203.156.266.109zm.656.656c.063-.063.031-.203-.063-.297-.109-.109-.25-.125-.313-.047-.078.063-.047.203.063.297.109.109.25.125.313.047zm.891.391c.031-.094-.063-.203-.203-.25-.125-.031-.266.016-.297.109s.063.203.203.234c.125.047.266 0 .297-.094zm.984.078c0-.109-.125-.187-.266-.172-.141 0-.25.078-.25.172 0 .109.109.187.266.172.141 0 .25-.078.25-.172zm.906-.156c-.016-.094-.141-.156-.281-.141-.141.031-.234.125-.219.234.016.094.141.156.281.125s.234-.125.219-.219z" /></svg>
-}
-
-export const LinkedIn = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 635 540"><path d="M32 .09C17.45 3.7 5.98 13.44 1.78 28-.79 36.91 0 51.38 0 61v432c.04 24.39 14.56 42.96 40 43h457c23.9-.04 39.72-17.51 40-41v-73V43C536.96 18.57 522.5.04 497 .09h-74-261-85-45zm80 75.21c11.04-1.1 20.12-.43 30 5.01 34.93 19.24 29.29 71.97-9 83.25-3.49 1.03-11.46 2.39-15 2.12-11.3-.87-22.89-5.52-30.83-13.77-19.53-20.3-16-54.53 7.83-69.82 5.58-3.58 10.51-5.57 17-6.79zM456 456h-78V318c-.03-20.3-5.51-44.46-27-51.9-5.83-2.02-10.94-2.17-17-2.1-4.8.06-9.38.89-14 2.16-23.26 6.4-29.53 28.57-31.09 49.84l-.91 12v128h-78V202h75v34c4.75-2.66 11.57-14.11 18-19.56 19.64-16.65 37.71-21.73 63-21.44 8.37.1 23.97 2.33 32 4.72 37.5 11.13 51.12 40.19 56.13 76.28l1.91 19v15l.96 12v37l-1 13v84zM159 202v254H81V202h78zm431 264.46c-7.26 1.62-13.55 4.28-18.91 9.63-22.72 22.69-7.89 63.14 24.91 63.9 10.97.25 21.99-3.76 29.33-12.16 16.35-18.7 10.37-49.99-13.33-59.37-6.96-2.75-14.66-3.01-22-2zm3 4.84c6.04-.6 11.33-.18 17 2.18 25.68 10.67 25.41 47.73 0 58.64-3.78 1.63-9.89 3.01-14 2.7-28.85-2.19-41.13-37.75-18.99-56.49 4.97-4.2 9.73-5.85 15.99-7.03zm-9 11.7v40h6v-17c12.23.43 9.66 11.38 17.32 15.98 1.94 1.16 4.5.99 6.68 1.02-1.9-5.97-7.14-13-11-18 13.31-1.07 11.27-15.58 6.77-19.35-4.8-4.03-19.33-2.65-25.77-2.65zm6 17v-12c3.62 0 9.77-.4 12.89 1.31 4.19 2.3 4.74 7.14.88 9.52-2.7 1.66-10.38 1.17-13.77 1.17z" /></svg>
-}
-
-export const Qiita = () => {
-  return <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 256 256'><path d='M82 .22c-16.13 3.2-28.59 7.83-42 17.52-9.75 7.05-18.97 16.9-25.14 27.26C-12.67 91.27.77 150.39 47 178.94c7.94 4.9 18.89 9.35 28 11.4 12.8 2.86 25.06 3.27 38 1.22 12.11-1.91 21.24-5.63 32-11.34 8.61-4.58 14.17-9.37 21-16.22 45.92-46.07 33.44-125.02-25-154.25-11.78-5.88-23.73-9.73-37-9.53H82zM202 0c8.12 11.71 15.54 19.7 22.19 33 14.31 28.62 18.26 61.85 11.39 93-2.99 13.53-7.27 25.78-13.89 38-1.88 3.48-9.36 13.75-8.82 17 .66 3.95 9.05 10.39 12.13 13.17 9.75 8.82 20.4 21.24 31 28.83V0h-54zM0 201v55h223c-6.76-9.45-22.28-22.3-31-31-2.61-2.6-9.63-10.58-13-10.99-2.52-.3-5.91 2.27-8 3.57-5.46 3.36-11.1 6.47-17 8.99-17.29 7.4-36.15 11.99-55 12.42-4.87.11-6.5-.39-11-.82-23.86-2.31-43.17-7.05-64-19.62L0 201zm255 24l-29 31h30l-1-31z' /></svg>
-}
-
-export const Twitter = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 400 400"><rect width="400" height="400" fill="none" /><path d="M153.62,301.59c94.34,0,145.94-78.16,145.94-145.94,0-2.22,0-4.43-.15-6.63A104.36,104.36,0,0,0,325,122.47a102.38,102.38,0,0,1-29.46,8.07,51.47,51.47,0,0,0,22.55-28.37,102.79,102.79,0,0,1-32.57,12.45,51.34,51.34,0,0,0-87.41,46.78A145.62,145.62,0,0,1,92.4,107.81a51.33,51.33,0,0,0,15.88,68.47A50.91,50.91,0,0,1,85,169.86c0,.21,0,.43,0,.65a51.31,51.31,0,0,0,41.15,50.28,51.21,51.21,0,0,1-23.16.88,51.35,51.35,0,0,0,47.92,35.62,102.92,102.92,0,0,1-63.7,22A104.41,104.41,0,0,1,75,278.55a145.21,145.21,0,0,0,78.62,23" /></svg>
-}
-
-export const Wantedly = () => {
-  return <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 900 900'><path d='M100 208.6h100l150 362.1L400 450 300 208.6h100l50 120.7 50-120.7h100L500 450l50 120.7 150-362.1h100L600 691.4H500l-50-120.7-50 120.7H300z' /></svg>
-}
-
-export const Hatena = () => {
-  return <svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'><path d='M20.47 0C22.42 0 24 1.58 24 3.53v16.94c0 1.95-1.58 3.53-3.53 3.53H3.53C1.58 24 0 22.42 0 20.47V3.53C0 1.58 1.58 0 3.53 0h16.94zm-3.705 14.47c-.78 0-1.41.63-1.41 1.41s.63 1.414 1.41 1.414 1.41-.645 1.41-1.425-.63-1.41-1.41-1.41zM8.61 17.247c1.2 0 2.056-.042 2.58-.12.526-.084.976-.222 1.32-.412.45-.232.78-.564 1.02-.99s.36-.915.36-1.48c0-.78-.21-1.403-.63-1.87-.42-.48-.99-.734-1.74-.794.66-.18 1.156-.45 1.456-.81.315-.344.465-.824.465-1.424 0-.48-.103-.885-.3-1.26-.21-.36-.493-.645-.883-.87-.345-.195-.735-.315-1.215-.405-.464-.074-1.29-.12-2.474-.12H5.654v10.486H8.61zm.736-4.185c.705 0 1.185.088 1.44.262.27.18.39.495.39.93 0 .405-.135.69-.42.855-.27.18-.765.254-1.44.254H8.31v-2.297h1.05zm8.656.706v-7.06h-2.46v7.06H18zM8.925 9.08c.71 0 1.185.08 1.432.24.245.16.367.435.367.83 0 .38-.13.646-.39.804-.265.154-.747.232-1.452.232h-.57V9.08h.615z' /></svg>
-}
-
-// General Website SVG
-
-export const Home = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48"><polyline points="4 23 24 6 44 23" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /><polyline points="9 26 9 42 39 42 39 26" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /><polyline points="39 12 39 6 34 6" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /></svg>
-}
-
-export const Human = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48"><rect width="48" height="48" fill="none" /><circle cx="24" cy="13" r="7" fill="none" stroke="#FFF" strokeMiterlimit="10" strokeWidth="3" /><path d="M10,44V38A10,10,0,0,1,20,28h8A10,10,0,0,1,38,38v6" fill="none" stroke="#FFF" strokeMiterlimit="10" strokeWidth="3" strokeLinecap="round" /></svg>
-}
-
-export const Cv = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" opacity="1"><path fill="#FFF" d="M378.418 0H195.122l-9.319 9.301L57.001 138.102l-9.301 9.309V426.1c0 47.357 38.534 85.9 85.9 85.9h244.819c47.348 0 85.882-38.542 85.882-85.9V85.882C464.3 38.524 425.767 0 378.418 0zm54.079 426.1c0 29.875-24.212 54.096-54.078 54.096H133.6c-29.875 0-54.096-24.221-54.096-54.096V160.587h83.719c24.882 0 45.073-20.183 45.073-45.074V31.802h170.123c29.866 0 54.078 24.222 54.078 54.079V426.1z" /><path fill="#FFF" d="M342.716 159.423c-7.668 5.367-17.004 8.537-27.052 8.537-10.056 0-19.384-3.17-27.061-8.537-13.418 5.775-21.66 15.902-26.688 24.847-6.679 11.864-1.442 28.659 10.092 28.659h87.306c11.542 0 16.787-16.795 10.1-28.659-5.029-8.946-13.271-19.072-26.697-24.847zM315.663 156.722c19.723 0 35.694-15.971 35.694-35.685v-8.554c0-19.696-15.971-35.694-35.694-35.694-19.714 0-35.702 15.998-35.702 35.694v8.554c0 19.714 15.989 35.685 35.702 35.685zM139.635 260.373h230.394v22.371H139.635zM139.635 327.488h230.394v22.38H139.635zM139.635 394.61h161.281v22.363H139.635z" /></svg>
-}
-
-export const Page = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path d="M0 48v416c0 26.508 21.492 48 48 48h416c26.508 0 48-21.492 48-48V48c0-26.508-21.492-48-48-48H48C21.492 0 0 21.492 0 48zm86.336 6c0 10.492-8.508 19-19 19s-19-8.508-19-19 8.508-19 19-19 19 8.508 19 19zm70.5 0c0 10.492-8.508 19-19 19s-19-8.508-19-19 8.508-19 19-19 19 8.508 19 19zm70.5 0c0 10.492-8.508 19-19 19s-19-8.508-19-19 8.508-19 19-19 19 8.508 19 19zM40 104h432v360c0 4.406-3.586 8-8 8H48c-4.414 0-8-3.594-8-8V104z" fill="#FFF" /><path d="M264 192h152v32H264zM88 352h328v32H88zM88 192h120v120H88zM282.958 304H264v-32h152v32H298.958z" fill="#FFF" /></svg>
-}
-
-export const Pen = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48" ><rect width="48" height="48" fill="none" /><circle cx="20" cy="28" r="4" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /><path d="M38.42,23.72l-5,12.79C26.13,36.51,4,44,4,44s7.49-23.21,7.49-29.46l12.79-5" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /><line x1="4" y1="44" x2="16.28" y2="31.72" fill="none" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /><rect x="30.41" y="1.33" width="7.76" height="24.34" transform="translate(0.5 28.2) rotate(-45)" strokeWidth="3" stroke="#FFF" strokeLinecap="round" strokeLinejoin="round" fill="none" /></svg>
-}
-
-// svgs from material design
-export const EditSvg = (props) => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill={props.fill} width={props.size} height={props.size}><path d="M0 0h24v24H0z" fill="none" /><path d="M3 17.25V21h3.75L17.81 9.94l-3.75-3.75L3 17.25zM20.71 7.04c.39-.39.39-1.02 0-1.41l-2.34-2.34c-.39-.39-1.02-.39-1.41 0l-1.83 1.83 3.75 3.75 1.83-1.83z" /></svg>
-}
-
-export const DashboardSvg = (props) => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill={props.fill} width={props.size} height={props.size}><path d="M0 0h24v24H0z" fill="none" /><path d="M3 13h8V3H3v10zm0 8h8v-6H3v6zm10 0h8V11h-8v10zm0-18v6h8V3h-8z" /></svg>
-}
-
-export const TagSvg = (props) => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill={props.fill} width={props.size} height={props.size}><path d="M0 0h24v24H0z" fill="none" /><path d="M21.41 11.58l-9-9C12.05 2.22 11.55 2 11 2H4c-1.1 0-2 .9-2 2v7c0 .55.22 1.05.59 1.42l9 9c.36.36.86.58 1.41.58.55 0 1.05-.22 1.41-.59l7-7c.37-.36.59-.86.59-1.41 0-.55-.23-1.06-.59-1.42zM5.5 7C4.67 7 4 6.33 4 5.5S4.67 4 5.5 4 7 4.67 7 5.5 6.33 7 5.5 7z" /></svg>
-}
-
-export const SearchSvg = (props) => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill={props.fill} width={props.size} height={props.size}><path d="M0 0h24v24H0z" fill="none" /><path d="M15.5 14h-.79l-.28-.27C15.41 12.59 16 11.11 16 9.5 16 5.91 13.09 3 9.5 3S3 5.91 3 9.5 5.91 16 9.5 16c1.61 0 3.09-.59 4.23-1.57l.27.28v.79l5 4.99L20.49 19l-4.99-5zm-6 0C7.01 14 5 11.99 5 9.5S7.01 5 9.5 5 14 7.01 14 9.5 11.99 14 9.5 14z" /></svg>
-}
-
-export const PersonSvg = (props) => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill={props.fill} width={props.size} height={props.sieze}><path d="M0 0h24v24H0z" fill="none" /><path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z" /></svg>
-}
-
-// Symbol Icon
-export const Arrow = () => {
-  return <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 48 48"><rect width="48" height="48" fill="none" /><polyline points="14.54 41.46 12 44 22 44 42 24 22 4 12 4 32 24 19.26 36.74" fill="#50CAF9" stroke="#50CAF9" strokeLinecap="round" strokeLinejoin="round" strokeWidth="3" /></svg>
-}
-
